cars-model

+++ data/migrations/01-make_cars_table.js
@@ -4,7 +4,7 @@ exports.up = function (knex) {
     tbl.string("vin", 17).notNullable().unique();
     tbl.string("make", 128).notNullable();
     tbl.string("model", 128).notNullable();
-    tbl.numeric("mileage").notNullable()
+    tbl.integer("mileage").notNullable()
     tbl.string("title", 128);
     tbl.string("transmission", 128)
   });
